hw8: adding ScrapperQueueConsumer, checkstyle fixes and some refactoring
