+++ bot/src/main/java/edu/java/bot/api/controller/http/BotController.java
@@ -1,11 +1,10 @@
-package edu.java.bot.api.controller;
+package edu.java.bot.api.controller.http;
 
 import edu.java.bot.api.service.BotService;
 import edu.java.model.LinkUpdate;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
-import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -18,8 +17,8 @@ public class BotController {
     private final BotService botService;
 
     @PostMapping
-    public ResponseEntity<Void> postUpdate(@RequestBody @Valid LinkUpdate req, BindingResult errors) {
-        botService.postUpdate(req, errors);
+    public ResponseEntity<Void> postUpdate(@RequestBody @Valid LinkUpdate req) {
+        botService.postUpdate(req);
         return ResponseEntity
             .ok()
             .build();
