+++ bot/src/main/java/edu/java/bot/configuration/bot/BotConfig.java
@@ -8,9 +8,8 @@ import edu.java.bot.command.ListCommand;
 import edu.java.bot.command.StartCommand;
 import edu.java.bot.command.TrackCommand;
 import edu.java.bot.command.UntrackCommand;
-import java.util.List;
-
 import edu.java.bot.configuration.ApplicationConfig;
+import java.util.List;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
