+++ scrapper/src/main/java/edu/java/scheduler/updater/NotificationSender.java
@@ -1,7 +1,7 @@
 package edu.java.scheduler.updater;
 
 import edu.java.client.BotClient.BotClient;
-import edu.java.client.ScrapperQueueProducer;
+import edu.java.client.Kafka.ScrapperQueueProducer;
 import edu.java.configuration.ApplicationConfig;
 import edu.java.model.LinkUpdate;
 import lombok.RequiredArgsConstructor;
@@ -17,15 +17,14 @@ public class NotificationSender {
     Не совсем понимаю задумку в виде написания отдельного сервиса.
     Как-будто бы тут просится просто создать конфиги по типу AccessType в дз6, где мы
     в зависимости от того, что указано в .yml будем инициализировать или HTTP-клиент, или
-    же Kafk'у. Я сделал, как требовалось в задании, а тут просто решил описать свои мысли,
-    чтобы получить какой-то комментарий насчёт своих мыслей.
+    же Kafk'у. Я сделал, как требовалось в задании, а тут просто решил описать свои мысли.
      */
 
     public void sendUpdate(LinkUpdate linkUpdate) {
         if (applicationConfig.useQueue()) {
             scrapperQueueProducer.send(linkUpdate);
         } else {
-            botClient.postUpdates(linkUpdate);
+            botClient.retryPostUpdates(linkUpdate);
         }
     }
 }
